Chore: Booking01 스타일, 타입 파일 정리
