--- src/pages/Booking/Booking.styled.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const BookingWrapper = styled.div`
-  width: 100%;
-  /* height: 100%; */
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 12px;
-`
