+++ src/pages/Booking/Booking.tsx
@@ -1,5 +1,5 @@
+import styled from 'styled-components'
 import CenterPageLayout from '../../components/Layouts/CenterPageLayout'
-import { BookingWrapper } from './Booking.styled'
 import Booking01 from './Booking01/Booking01'
 import Booking02 from './Booking02/Booking02'
 import { useSearchParams } from 'react-router-dom'
@@ -16,3 +16,13 @@ const Booking = () => {
 }
 
 export default Booking
+
+const BookingWrapper = styled.div`
+  width: 100%;
+  /* height: 100%; */
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 12px;
+`
