+++ src/pages/Booking/Booking01/ui/SelectBody/ui/ErrorMessage.tsx
@@ -1,12 +1,8 @@
-import React from 'react'
-import { MsgBoxStyle } from '../SelectBody.style'
-
-interface IErrorMessageProps {
-  errorMsg: React.ReactNode
-}
+import BS1 from '../../../Booking01.styled'
+import { IErrorMessageProps } from '../../../Booking01.types'
 
 const ErrorMessage = ({ errorMsg }: IErrorMessageProps) => {
-  return <MsgBoxStyle>{errorMsg}</MsgBoxStyle>
+  return <BS1.MsgBoxStyle>{errorMsg}</BS1.MsgBoxStyle>
 }
 
 export default ErrorMessage
