+++ src/pages/Booking/Booking01/ui/SelectBody/ui/SelectTitle.tsx
@@ -1,11 +1,8 @@
-import { SelectTitleStyle } from '../SelectBody.style'
-
-interface ISelectTitleProps {
-  title?: string
-}
+import BS1 from '../../../Booking01.styled'
+import { ISelectTitleProps } from '../../../Booking01.types'
 
 const SelectTitle = ({ title = '타이틀' }: ISelectTitleProps) => {
-  return <SelectTitleStyle>{title}</SelectTitleStyle>
+  return <BS1.SelectTitleStyle>{title}</BS1.SelectTitleStyle>
 }
 
 export default SelectTitle
