--- src/pages/Booking/Booking01/ui/SelectDateCarousel/SelectDateCarousel.types.ts
@@ -1,10 +1,0 @@
-export interface IArrowButtonProps {
-  onClick?: React.MouseEventHandler<HTMLButtonElement>
-}
-
-export interface IDateButtonProps {
-  dateNumber: string
-  today: string
-  isSelected: boolean
-  setIsSelected: any
-}
