--- src/pages/Booking/Booking01/ui/SelectDateCarousel/ui/NextButton.tsx
@@ -1,13 +1,0 @@
-import { WeuiArrowOutlined } from '../../../../../../assets/svg/WeuiArrowOutlined'
-import { NextBtn } from '../SelectDateCarousel.styled'
-import { IArrowButtonProps } from '../SelectDateCarousel.types'
-
-const NextButton = ({ onClick }: IArrowButtonProps) => {
-  return (
-    <NextBtn onClick={onClick}>
-      <WeuiArrowOutlined />
-    </NextBtn>
-  )
-}
-
-export default NextButton
