--- src/pages/Booking/Booking01/ui/SelectDateCarousel/ui/PrevButton.tsx
@@ -1,13 +1,0 @@
-import { WeuiArrowOutlined } from '../../../../../../assets/svg/WeuiArrowOutlined'
-import { PrevBtn } from '../SelectDateCarousel.styled'
-import { IArrowButtonProps } from '../SelectDateCarousel.types'
-
-const PrevButton = ({ onClick }: IArrowButtonProps) => {
-  return (
-    <PrevBtn onClick={onClick}>
-      <WeuiArrowOutlined />
-    </PrevBtn>
-  )
-}
-
-export default PrevButton
