+++ src/pages/Booking/Booking02/ui/SelectSeatBody/ui/SeatStatus.tsx
@@ -1,7 +1,7 @@
 import SeatUnchecked from '../../../../../../assets/svg/seat_unchecked.svg'
 import SeatChecked from '../../../../../../assets/svg/seat_checked.svg'
 import SeatDisabled from '../../../../../../assets/svg/seat_disabled.svg'
-import BS2 from '../../Booking02.styled'
+import BS2 from '../../../Booking02.styled'
 
 const SeatStatus = () => {
   return (
